\name{derivErfc}
\alias{derivErfc}

\title{Calculate High Order Derivatives of Composite Functions}

\description{
    Symbolically calculates high order derivative of composite functions 
    involving complementary error function and exponential function. 
}

\references{

  \cite{Comtet, L (1974) Advanced combinatorics: {T}he art of finite and infinite 
    expansions. Boston: D. Reidel Pub.}  
  
  \cite{Hankin, R.K.S (2006) {A}dditive integer partitions in {R}. 
    Journal of Statistical Software, Code Snippets, \bold{16}.}
    
  \cite{Olver, F.W.J., Lozier, D.W., Boisver, R.F., and Clark,
  C.W (2010) Handbook of Mathematical Functions. 
  New York: National Institute of Standards and Technology, 
  and Cambridge University Press.}     
}

\note{
    A C++ version of this function will be added in due 
    course. It involves very intensive combinatorial computation 
    and is not supposed to be called by the user. 	
}

\author{
  Thanh T. Tran \email{frmqa.package@gmail.com}
}        


