\name{frmqa-package}
\alias{frmqa-package}
\alias{frmqa}
\docType{package}
\title{
A Package for \bold{F}inancial \bold{R}isk Management 
and \bold{Q}uantitative \bold{A}nalysis
}
\description{
This package comprises of R and C++ functions which deal with 
issues relating 
to financial risk management and quantitative analysis by applying 
uni- and multi-variate generalized hyperbolic and related distributions. 
These issues are approached from both directions: analytical (i.e., deriving 
and programing analytic formulae) and numerical. Note that the latter 
appears to be the only approach currently used in practice because of the 
intractability of these families of distributions, which is caused by 
the presence of the modified Bessel function of the third kind 
BesselK \eqn{K_{\lambda}(z)}, in their density functions. In this 
package, the naming of special functions and their related 
functions (e.g., BesselK and incomplete BesselK functions) 
follows the convention used in the R 
package \bold{gsl}. 
}
\details{
\tabular{ll}{
Package: \tab frmqa\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2012-11-2\cr
License: \tab GPL (>= 2)\cr
}
The package is under current development with more 
functions are planned to be added shortly.
}
