\name{frmqa-package}
\alias{frmqa-package}
\alias{frmqa}
\docType{package}
\title{
A Package for \bold{F}inancial \bold{R}isk Management and \bold{Q}uantitative \bold{A}nalysis
}
\description{
This package comprises of R and C++ functions which focus on dealing with issues relating to financial 
risk management and quantitative analysis applying uni- and multi-variate generalized hyperbolic and 
related distributions. These issues are approached from both directions: analytical (i.e., deriving 
and programing analytic formulae) and numerical. In this package, the naming of special functions and their 
related functions (e.g., the modified Bessel function of the third kind BesselK and incomplete BesselK functions) follows the convention used in the R 
package \bold{gsl}. 
}
\details{
\tabular{ll}{
Package: \tab frmqa\cr
Type: \tab Package\cr
Version: \tab 0.1-3\cr
Date: \tab 2012-07-29\cr
License: \tab GPL (>= 2)\cr
}
Note that the package is under current development with more functions are planned to be added shortly.
}
